+++ src/main/java/com/codetaylor/mc/artisanworktables/proxy/ClientProxy.java
@@ -1,0 +1,13 @@
+package com.codetaylor.mc.artisanworktables.proxy;
+
+import net.minecraft.client.Minecraft;
+
+public class ClientProxy
+    implements IProxy {
+
+  @Override
+  public boolean isIntegratedServerRunning() {
+
+    return Minecraft.getMinecraft().isIntegratedServerRunning();
+  }
+}