+++ src/main/java/com/codetaylor/mc/artisanworktables/proxy/ServerProxy.java
@@ -1,0 +1,11 @@
+package com.codetaylor.mc.artisanworktables.proxy;
+
+public class ServerProxy
+    implements IProxy {
+
+  @Override
+  public boolean isIntegratedServerRunning() {
+
+    return false;
+  }
+}